+++ examples/todoapp/integration_test/tests/add_todo_page_test.dart
@@ -1,4 +1,4 @@
-import '../main/base/test_constants.dart';
+import '../main/base/tags.dart';
 import '../main/configuration/build_app.dart';
 import '../main/configuration/patrol_base_config.dart';
 import '../main/steps_utils/todo_list_stepts.dart';
@@ -8,9 +8,7 @@ void main() {
   final patrolBaseConfig = PatrolBaseConfig();
 
   patrolBaseConfig.patrol(
-    '''
-    Add a todo item with title and description to todo list,
-    ''',
+    'Add a todo item with title and description to todo list',
     ($) async {
       BuildApp app = BuildApp($);
       await app.buildApp();