+++ packages/rx_bloc_cli/mason_templates/bricks/feature_login/__brick__/test/feature_login/view/login_golden_test.dart
@@ -1,0 +1,31 @@
+import 'package:{{project_name}}/base/models/errors/error_model.dart';
+
+import '../../helpers/golden_helper.dart';
+import '../../helpers/models/scenario.dart';
+import '../factory/login_factory.dart';
+import '../stubs.dart';
+
+void main() {
+  runGoldenTests([
+    generateDeviceBuilder(
+        widget:
+        loginFactory(loggedIn: false, showErrors: false, isLoading: false),
+        scenario: Scenario(name: 'login_empty')),
+    generateDeviceBuilder(
+        widget: loginFactory(email: Stubs.email, password: Stubs.password),
+        scenario: Scenario(name: 'login_filled')),
+    generateDeviceBuilder(
+        widget:
+        loginFactory(isLoading: false, loggedIn: true, showErrors: false),
+        scenario: Scenario(name: 'login_success')),
+    generateDeviceBuilder(
+        widget: loginFactory(isLoading: true),
+        scenario: Scenario(name: 'login_loading')),
+    generateDeviceBuilder(
+        widget: loginFactory(
+            showErrors: true,
+            errors: UnknownErrorModel(
+                exception: Exception('Something went wrong'))),
+        scenario: Scenario(name: 'login_error'))
+  ]);
+}