+++ packages/rx_bloc_cli/mason_templates/bricks/lib_auth/__brick__/test/lib_auth/mock/auth_service_mock.dart
@@ -1,0 +1,9 @@
+import 'package:mockito/annotations.dart';
+import 'package:{{project_name}}/lib_auth/services/auth_service.dart';
+
+import 'auth_service_mock.mocks.dart';
+
+@GenerateMocks([
+  AuthService,
+])
+AuthService authServiceMockFactory() => MockAuthService();