+++ packages/rx_bloc_cli/mason_templates/bricks/lib_dev_menu/__brick__/lib/l10n/network_lookup/lib_dev_menu_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nLibDevMenuLookup extends I18nLibDevMenuLookup {
           bundledI18n = I18nLibDevMenu(I18nLibDevMenuLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nLibDevMenu(I18nLibDevMenuLookup_en());
-          break;
         default:
           bundledI18n = I18nLibDevMenu(I18nLibDevMenuLookup_en());
           break;