+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/lib/lib_pin_code/views/create_pin_page.dart
@@ -56,6 +56,8 @@ class _CreatePinPageState extends State<CreatePinPage> {
                   child: PinCodeKeyboard(
                     mapBiometricMessageToString: (message) =>
                         _exampleMapMessageToString(message, context),
+                    biometricsAuthDataSource:
+                        context.read<PinBiometricsAuthDataSource?>(),
                     pinCodeService: context.read<CreatePinCodeService>(),
                     translateError: (error) =>
                         error.asErrorModel().translate(context),