+++ packages/rx_bloc_cli/mason_templates/bricks/lib_pin_code/__brick__/test/lib_pin_code/stubs.dart
@@ -1,0 +1,11 @@
+import 'package:local_session_timeout/local_session_timeout.dart';
+
+class Stubs {
+  static const pin = '2134';
+
+  static const title = 'Some title';
+
+  static const sessionStartListening = SessionState.startListening;
+
+  static const sessionStopListening = SessionState.stopListening;
+}