+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/l10n/network_lookup/error_lookup.dart
@@ -21,8 +21,6 @@ class AppI18nErrorLookup extends I18nErrorLookup {
           bundledI18n = I18nError(I18nErrorLookup_bg());
           break;
         case 'en':
-          bundledI18n = I18nError(I18nErrorLookup_en());
-          break;
         default:
           bundledI18n = I18nError(I18nErrorLookup_en());
           break;