+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/test/base/common_ui_components/view/mocks/app_error_modal_widget_mock.dart
@@ -1,0 +1,11 @@
+import 'package:mockito/annotations.dart';
+import 'package:rx_bloc/rx_bloc.dart';
+
+import 'app_error_modal_widget_mock.mocks.dart';
+
+@GenerateMocks([RxBlocTypeBase])
+RxBlocTypeBase appErrorModalWidgetMockFactory() {
+  final blocMock = MockRxBlocTypeBase();
+
+  return blocMock;
+}