+++ packages/rx_bloc_cli/test/models/configurations/feature_configuration_test.dart
@@ -10,12 +10,9 @@ void main() {
         'analytics or push notifications are enabled', () {
       sut = FeatureConfiguration(
         changeLanguageEnabled: false,
-        counterEnabled: false,
-        widgetToolkitEnabled: false,
         analyticsEnabled: false,
         pushNotificationsEnabled: true,
         realtimeCommunicationEnabled: false,
-        deepLinkEnabled: false,
         devMenuEnabled: false,
         patrolTestsEnabled: false,
         cicdEnabled: false,
@@ -27,12 +24,9 @@ void main() {
 
       sut = FeatureConfiguration(
         changeLanguageEnabled: false,
-        counterEnabled: false,
-        widgetToolkitEnabled: false,
         analyticsEnabled: true,
         pushNotificationsEnabled: false,
         realtimeCommunicationEnabled: false,
-        deepLinkEnabled: false,
         devMenuEnabled: false,
         patrolTestsEnabled: false,
         cicdEnabled: false,
@@ -44,12 +38,9 @@ void main() {
 
       sut = FeatureConfiguration(
         changeLanguageEnabled: false,
-        counterEnabled: false,
-        widgetToolkitEnabled: false,
         analyticsEnabled: true,
         pushNotificationsEnabled: true,
         realtimeCommunicationEnabled: false,
-        deepLinkEnabled: false,
         devMenuEnabled: false,
         patrolTestsEnabled: false,
         cicdEnabled: false,
@@ -61,12 +52,9 @@ void main() {
 
       sut = FeatureConfiguration(
         changeLanguageEnabled: false,
-        counterEnabled: false,
-        widgetToolkitEnabled: false,
         analyticsEnabled: false,
         pushNotificationsEnabled: false,
         realtimeCommunicationEnabled: false,
-        deepLinkEnabled: false,
         devMenuEnabled: false,
         patrolTestsEnabled: false,
         cicdEnabled: false,