+++ packages/rx_bloc_cli/test/stub.dart
@@ -5,6 +5,7 @@ import 'package:rx_bloc_cli/src/models/command_arguments/create_command_argument
 import 'package:rx_bloc_cli/src/models/configurations/auth_configuration.dart';
 import 'package:rx_bloc_cli/src/models/configurations/feature_configuration.dart';
 import 'package:rx_bloc_cli/src/models/configurations/project_configuration.dart';
+import 'package:rx_bloc_cli/src/models/configurations/showcase_configuration.dart';
 import 'package:rx_bloc_cli/src/models/generator_arguments.dart';
 import 'package:rx_bloc_cli/src/models/realtime_communication_type.dart';
 
@@ -26,6 +27,19 @@ final class Stub {
   static const incorrectCICDTypeCase = 'incorrect_case';
 
   static const licenseFileName = '{{~ licence.dart }}';
+
+  static const overwrittenLoginFiles = [
+    'test/feature_login/view/goldens/dark_theme/login_empty.png',
+    'test/feature_login/view/goldens/dark_theme/login_error.png',
+    'test/feature_login/view/goldens/dark_theme/login_filled.png',
+    'test/feature_login/view/goldens/dark_theme/login_loading.png',
+    'test/feature_login/view/goldens/dark_theme/login_success.png',
+    'test/feature_login/view/goldens/light_theme/login_empty.png',
+    'test/feature_login/view/goldens/light_theme/login_error.png',
+    'test/feature_login/view/goldens/light_theme/login_filled.png',
+    'test/feature_login/view/goldens/light_theme/login_loading.png',
+    'test/feature_login/view/goldens/light_theme/login_success.png'
+  ];
 
   static Map<String, Object> get defaultValues {
     var map = <String, Object>{
@@ -91,16 +105,21 @@ final class Stub {
       analyticsEnabled: true,
       pushNotificationsEnabled: true,
       changeLanguageEnabled: true,
-      counterEnabled: true,
       patrolTestsEnabled: true,
       devMenuEnabled: true,
-      deepLinkEnabled: true,
-      widgetToolkitEnabled: true,
       realtimeCommunicationEnabled: true,
       cicdEnabled: true,
       cicdGithubEnabled: true,
       cicdCodemagicEnabled: true,
       profileEnabled: true,
+    ),
+    showcaseConfiguration: ShowcaseConfiguration(
+      counterEnabled: true,
+      widgetToolkitEnabled: true,
+      qrScannerEnabled: true,
+      deepLinkEnabled: true,
+      mfaEnabled: true,
+      otpEnabled: true,
     ),
   );
 
@@ -122,11 +141,8 @@ final class Stub {
       pushNotificationsEnabled: true,
       changeLanguageEnabled:
           CreateCommandArguments.changeLanguage.defaultValue(),
-      counterEnabled: CreateCommandArguments.counter.defaultValue(),
       patrolTestsEnabled: CreateCommandArguments.patrol.defaultValue(),
       devMenuEnabled: CreateCommandArguments.devMenu.defaultValue(),
-      deepLinkEnabled: CreateCommandArguments.deepLink.defaultValue(),
-      widgetToolkitEnabled: CreateCommandArguments.widgetToolkit.defaultValue(),
       realtimeCommunicationEnabled:
           CreateCommandArguments.realtimeCommunication.defaultValue() !=
               RealtimeCommunicationType.none,
@@ -136,6 +152,14 @@ final class Stub {
       cicdCodemagicEnabled:
           CreateCommandArguments.cicd.defaultValue() == CICDType.codemagic,
       profileEnabled: CreateCommandArguments.profile.defaultValue(),
+    ),
+    showcaseConfiguration: ShowcaseConfiguration(
+      counterEnabled: CreateCommandArguments.counter.defaultValue(),
+      widgetToolkitEnabled: CreateCommandArguments.widgetToolkit.defaultValue(),
+      deepLinkEnabled: CreateCommandArguments.deepLink.defaultValue(),
+      qrScannerEnabled: CreateCommandArguments.qrScanner.defaultValue(),
+      mfaEnabled: CreateCommandArguments.mfa.defaultValue(),
+      otpEnabled: CreateCommandArguments.otp.defaultValue(),
     ),
   );
 }